refactor(journal): extract scenery rendering into InkScrollView+Scenery
The fade-resistance change tipped InkScrollView.swift over SwiftLint's
1000-line hard limit. The scenery section (sceneryForDot and its
weather-mood tint helper) is a coherent unit — move it to its own
extension file.
